Add docker image for running tests
This change adds a Dockerfile, that can be used to build a docker image
containing this project. Running the docker container allows to run
gatling tests using the git-protocol for a set of given data- and
scenario-files that have to be mounted into the container. The
configuration can be provided by setting the container's environmental
variables.

Change-Id: If557d5e96704816dc0389ddba73ea6fdbcd7ad45